convert session footer into component, reuse across all pages
